requirements.txt
Django
django-extensions
django-bootstrap3
django-admin-bootstrapped
django-registration
django-cors-headers

psycopg2
south

meld3
supervisor
wsgiref
gunicorn

pyyaml
pytz

coverage

djangorestframework
django-rest-swagger
markdown
django-filter

passlib==1.6.2
aniso8601
ipython

ftfy

django-oauth-toolkit
django-allauth

django-leaflet
django-geojson

Pillow
